--- a/data_rentgen/db/models/column_lineage.py
+++ b/data_rentgen/db/models/column_lineage.py
@@ -26,6 +26,7 @@ from data_rentgen.db.models.run import Run
 class ColumnLineage(Base):
     __tablename__ = "column_lineage"
     __table_args__ = (
+        # in most cases we filter rows by created_at, and never by id
         PrimaryKeyConstraint("created_at", "id"),
         {"postgresql_partition_by": "RANGE (created_at)"},
     )