--- a/docs/changelog/next_release/247.feature.rst
+++ b/docs/changelog/next_release/247.feature.rst
@@ -1,0 +1,1 @@
+Speedup fetching ``Run`` and ``Operation`` from database by id.