<!-- Thank you for your contribution! -->
<!-- Unless your change is trivial, please create an issue to discuss the change before creating a PR -->
<!-- See https://github.com/MobileTeleSystems/data-rentgen/blob/develop/CONTRIBUTING.rst for help on Contributing -->
<!-- PLEASE DO **NOT** put issue ids in the PR title! Instead, add a descriptive title and put ids in the body -->

## Change Summary

Tables `input`, `output`, `column_lineage` in most cases are selected with `created_at >=` filter, and no filters on `id` column. Because B-Tree indexes are prefix-based, `created_at` should be the first column in `PRIMARY KEY (created_at, id)`, and this is already true.

But for `run`, `operation` its actually the opposite - in most cases we filter runs or operations by id, and then by `created_at` (mostly to avoid scanning all partitions). So for these two tables we should use `PRIMARY KEY (id, created_at)`.

Before:
```sql
SELECT * FROM run
WHERE
    run.created_at >= '2025-03-21 08:50:06.115000+00:00'::TIMESTAMP WITH TIME ZONE
AND run.created_at <= '2025-03-29 09:20:38.781000+00:00'::TIMESTAMP WITH TIME ZONE
AND run.id IN (
  '0195b7ef-ed5c-7cc9-92a8-2cdc33cccfb5'::uuid,
  '0195c23c-4127-79a7-b8d2-43c42a786a49'::uuid,
  '0195b7ef-f073-739b-8c61-1be8b08891ac'::uuid,
  '0195d6d6-1324-723e-9e60-1076e610cd81'::uuid,
  '0195c23c-15ff-7f34-ae73-19caf630365a'::uuid,
  '0195bd16-920f-774c-933c-78df12d1f31b'::uuid
)
```

```
 Index Scan using run_y2025_m03_pkey on run_y2025_m03 run  (cost=0.29..792.59 rows=5 width=472) (actual time=0.123..13.065 rows=6 loops=1)
   Index Cond: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (created_at <= '2025-03-29 09:20:38.781+00'::timestamp with time zone) AND (id = ANY ('{0195b7ef-ed5c-7cc9-92a8-2cdc33cccfb5,0195c23c-4127-79a7-b8d2-43c42a786a49,0195b7ef-f073-739b-8c61-1be8b08891ac,0195d6d6-1324-723e-9e60-1076e610cd81,0195c23c-15ff-7f34-ae73-19caf630365a,0195bd16-920f-774c-933c-78df12d1f31b}'::uuid[])))
 Planning Time: 1.273 ms
 Execution Time: 13.129 ms
(4 rows)
```

After:
```
 Index Scan using run_y2025_m03_pkey on run_y2025_m03 run  (cost=0.29..39.63 rows=5 width=472) (actual time=0.034..0.071 rows=6 loops=1)
   Index Cond: ((id = ANY ('{0195b7ef-ed5c-7cc9-92a8-2cdc33cccfb5,0195c23c-4127-79a7-b8d2-43c42a786a49,0195b7ef-f073-739b-8c61-1be8b08891ac,0195d6d6-1324-723e-9e60-1076e610cd81,0195c23c-15ff-7f34-ae73-19caf630365a,0195bd16-920f-774c-933c-78df12d1f31b}'::uuid[])) AND (created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (created_at <= '2025-03-29 09:20:38.781+00'::timestamp with time zone))
 Planning Time: 1.538 ms
 Execution Time: 0.107 ms
(4 rows)
```


Note: For queries like:
* `SELECT run WHERE job_id = ... AND created_at >= ...`
* `SELECT operations WHERE run_id = ... AND created_at >= ...`

There are no filters on `id` column, so they can become slower. But `job_id` already has quite high selectivity, so the slowdown is not very profound.

<details>

```sql
SELECT * FROM run
WHERE
    run.created_at >= '2025-03-21 08:50:06.115000+00:00'::TIMESTAMP WITH TIME ZONE
AND run.job_id IN (1804, 632, 1073, 1271);
```

Before:
```
 Append  (cost=17.40..124.31 rows=31 width=470) (actual time=0.021..0.050 rows=36 loops=1)
   ->  Bitmap Heap Scan on run_y2025_m03 run_1  (cost=17.40..124.15 rows=27 width=472) (actual time=0.020..0.044 rows=36 loops=1)
         Recheck Cond: (job_id = ANY ('{1804,632,1073,1271}'::bigint[]))
         Filter: (created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone)
         Heap Blocks: exact=27
         ->  Bitmap Index Scan on run_y2025_m03_job_id_idx  (cost=0.00..17.39 rows=30 width=0) (actual time=0.014..0.014 rows=36 loops=1)
               Index Cond: (job_id = ANY ('{1804,632,1073,1271}'::bigint[]))
   ->  Seq Scan on run_y2025_m04 run_2  (cost=0.00..0.00 rows=1 width=462) (actual time=0.002..0.002 rows=0 loops=1)
         Filter: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (job_id = ANY ('{1804,632,1073,1271}'::bigint[])))
   ->  Seq Scan on run_y2025_m05 run_3  (cost=0.00..0.00 rows=1 width=462) (actual time=0.001..0.001 rows=0 loops=1)
         Filter: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (job_id = ANY ('{1804,632,1073,1271}'::bigint[])))
   ->  Seq Scan on run_y2025_m06 run_4  (cost=0.00..0.00 rows=1 width=462) (actual time=0.001..0.001 rows=0 loops=1)
         Filter: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (job_id = ANY ('{1804,632,1073,1271}'::bigint[])))
   ->  Seq Scan on run_y2025_m07 run_5  (cost=0.00..0.00 rows=1 width=462) (actual time=0.001..0.001 rows=0 loops=1)
         Filter: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (job_id = ANY ('{1804,632,1073,1271}'::bigint[])))
 Planning Time: 0.440 ms
 Execution Time: 0.069 ms
(17 rows)
```

After:
```
 Append  (cost=17.40..124.31 rows=31 width=470) (actual time=0.106..0.255 rows=36 loops=1)
   ->  Bitmap Heap Scan on run_y2025_m03 run_1  (cost=17.40..124.15 rows=27 width=472) (actual time=0.105..0.221 rows=36 loops=1)
         Recheck Cond: (job_id = ANY ('{1804,632,1073,1271}'::bigint[]))
         Filter: (created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone)
         Heap Blocks: exact=27
         ->  Bitmap Index Scan on run_y2025_m03_job_id_idx  (cost=0.00..17.39 rows=30 width=0) (actual time=0.068..0.069 rows=36 loops=1)
               Index Cond: (job_id = ANY ('{1804,632,1073,1271}'::bigint[]))
   ->  Seq Scan on run_y2025_m04 run_2  (cost=0.00..0.00 rows=1 width=462) (actual time=0.008..0.008 rows=0 loops=1)
         Filter: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (job_id = ANY ('{1804,632,1073,1271}'::bigint[])))
   ->  Seq Scan on run_y2025_m05 run_3  (cost=0.00..0.00 rows=1 width=462) (actual time=0.003..0.004 rows=0 loops=1)
         Filter: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (job_id = ANY ('{1804,632,1073,1271}'::bigint[])))
   ->  Seq Scan on run_y2025_m06 run_4  (cost=0.00..0.00 rows=1 width=462) (actual time=0.003..0.003 rows=0 loops=1)
         Filter: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (job_id = ANY ('{1804,632,1073,1271}'::bigint[])))
   ->  Seq Scan on run_y2025_m07 run_5  (cost=0.00..0.00 rows=1 width=462) (actual time=0.003..0.003 rows=0 loops=1)
         Filter: ((created_at >= '2025-03-21 08:50:06.115+00'::timestamp with time zone) AND (job_id = ANY ('{1804,632,1073,1271}'::bigint[])))
 Planning Time: 2.444 ms
 Execution Time: 0.375 ms
(17 rows)
```
</details>

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this change? -->
<!-- WARNING: please use "fix #123" style references so the issue is closed when this PR is merged. -->

## Checklist

* [X] Commit message and PR title is comprehensive
* [X] Keep the change as small as possible
* [ ] Unit and integration tests for the changes exist
* [X] Tests pass on CI and coverage does not decrease
* [ ] Documentation reflects the changes where applicable
* [ ] `docs/changelog/next_release/<pull request or issue id>.<change type>.rst` file added describing change
  (see [CONTRIBUTING.rst](https://github.com/MobileTeleSystems/data-rentgen/blob/develop/CONTRIBUTING.rst) for details.)
* [x] My PR is ready to review.